ImConcluido a publicaação no  Kafka

+++ pedidos/src/main/java/br/com/Icompras/pedidos/client/representation/ClienteRepresentation.java
@@ -1,11 +1,13 @@
 package br.com.Icompras.pedidos.client.representation;
 
-public record ClienteRepresentation(Integer codigo,
+public record ClienteRepresentation(Long codigo,
                                     String nome,
                                     String cpf,
                                     String logradouro,
                                     String numero,
                                     String bairro,
                                     String email,
-                                    String telefone) {
+                                    String telefone
+)
+{
 }
